Version 2
When creating the ConverterLogic and creating the Segue, I was having a hard time getting them to take in the proper values when the convert button is hit. I was able to get an alert to pop up and it doesn't display that the conversion are happening. If you comment out the alert, you can then see that modal Segue pop up when you click convert, but it will not display the proper values. I did all I could think of. I am sure it's probably an easy fix, but none the less I tried. :)